util: add logging API that's configurable in runtime
oc_log now has configurable level in both runtime and compile
time:

OC_LOG_LEVEL_DISABLED (-1)
OC_LOG_LEVEL_ERROR (3)
OC_LOG_LEVEL_WARNING (4)
OC_LOG_LEVEL_NOTICE (5)
OC_LOG_LEVEL_INFO (6)
OC_LOG_LEVEL_DEBUG (7)
OC_LOG_LEVEL_TRACE (8)

To enable logs of given level and higher at compile time set
OC_LOG_MAXIMUM_LEVEL to value of the maximal level that should
be allowed.

To change the runtime configuration use oc_log_set_level. The
function is thread-safe and the runtime log level can be changed
at any point of the program.

Examples:
- Calling oc_log_set_level(OC_LOG_LEVEL_NOTICE) will enable logs
with log levels: OC_LOG_LEVEL_ERROR, OC_LOG_LEVEL_WARNING
and OC_LOG_LEVEL_NOTICE and disable the rest.
- Setting OC_LOG_MAXIMUM_LEVEL=-1 will disable and remove all
logs during compilation.

+++ messaging/coap/coap.c
@@ -925,8 +925,8 @@ coap_oscore_parse_options(void *packet, const uint8_t *data, size_t data_len,
                           bool oscore, bool validate)
 {
   if (data_len > UINT32_MAX) {
-    OC_WRN("message size(%zu) exceeds limit for coap message(%u)", data_len,
-           UINT32_MAX);
+    OC_WRN("message size(%zu) exceeds limit for coap message(%lu)", data_len,
+           (long unsigned)UINT32_MAX);
     return BAD_REQUEST_4_00;
   }
   coap_packet_t *const coap_pkt = (coap_packet_t *)packet;
@@ -1357,7 +1357,7 @@ coap_udp_parse_message(void *packet, uint8_t *data, size_t data_len,
 {
   if (data_len > UINT16_MAX) {
     OC_WRN("message size(%zu) exceeds limit for UDP message(%u)", data_len,
-           UINT16_MAX);
+           (unsigned)UINT16_MAX);
     return BAD_REQUEST_4_00;
   }
   coap_packet_t *const coap_pkt = (coap_packet_t *)packet;
@@ -1428,8 +1428,8 @@ coap_tcp_parse_message(void *packet, uint8_t *data, size_t data_len,
                        bool validate)
 {
   if (data_len > UINT32_MAX) {
-    OC_WRN("message size(%zu) exceeds limit for TCP message(%u)", data_len,
-           UINT32_MAX);
+    OC_WRN("message size(%zu) exceeds limit for TCP message(%lu)", data_len,
+           (long unsigned)UINT32_MAX);
     return BAD_REQUEST_4_00;
   }
   coap_packet_t *const coap_pkt = (coap_packet_t *)packet;
